Successsfully fixed hashing import
How ever now I have this wonderful runtime error. The stack overflows for main. It is beacuse I am returning the variable of the sha256 has back to the main function. I will have to to move the String to the heap to stop that from happening. However I have literally no idea how to do that but this is progress.